+++ src/Clusters.h
@@ -14,51 +14,48 @@
 #include <string>
 #include <numeric>
 #include <assert.h>
+#include <math.h>
 
 
 using namespace std;
 
-const int DEBUG = 0;
+const int DEBUG = 1;
 
 struct grid {
     vector<vector<double> > lines;
     vector<int> num_lines;
 };
 
-struct cluster {
-    vector<double[2]> *points;
-    double *centers[2];
-    int size = 0;
-};
+//struct cluster {
+//    vector<double[2]> *points;
+//    double *centers[2];
+//    int size = 0;
+//};
 
 class Cluster {
 private:
     int num_clusters = 0;
-
-    // < clusters < dims > >
-//    vector<cluster> clusters;
-
-    // < clusters < dims < points > > >
     vector<vector<vector<double> > > cluster_points;
+    //vector<vector<double> > cluster_centers;
+    vector<vector<double> > cluster_medians;
 
-    vector<vector<double> > cluster_centers;
     grid grids;
+
+    Cluster(vector<vector<double> > medians);
+
+    Cluster(vector<vector<double> > medians, vector<vector<vector<double> > > data);
 public:
-    Cluster(vector<vector<double> > centers);
+    Cluster(vector<int> labels, vector<vector<double> > medians, vector<vector<double> > data);
 
-    Cluster(vector<vector<double> > centers, vector<vector<vector<double> > > data);
-
-    Cluster(vector<int> labels, vector<vector<double> > centers, vector<vector<double>> data);
-
-    void set_clusters(vector<int> labels, vector<vector<double> > centers, vector<vector<double> > data);
+    Cluster(int k, vector<int> labels, vector<vector<double> > data);
 
     vector<vector<vector<double> > > get_points();
 
-    vector<vector<double> > get_centers();
-
     vector<vector<double> > get_points(int index);
 
-    vector<double> get_centers(int index);
+    vector<vector<double> > get_medians();
+
+    vector<double> get_medians(int index);
 
     void set_grids(grid G);